REFACTOR(MainActivity): Moved connection and DataStore test methods to utils
