+++ app/src/main/java/com/copay/app/MainActivity.kt
@@ -43,54 +43,13 @@ class MainActivity : ComponentActivity() {
         }
 
         // Method to verify the connection with the backend.
-        testApiConnection()
+        ConnectionManager.testApiConnection()
 
         // Method to verify the connection with the backend.
-        testDataStore()
-    }
-
-    private fun testApiConnection() {
-        // Using CoroutineScope to handle the asynchronous call on the IO thread.
-        CoroutineScope(Dispatchers.IO).launch {
-            try {
-                val response = ConnectionManager.testConnection()
-
-                // Returning to the main thread to log the result.
-                withContext(Dispatchers.Main) {
-
-                    if (response != null) {
-                        Log.d("TestConnection", "Connection Successful")
-                    } else {
-                        Log.e("TestConnection", "Error: Response body is null")
-                    }
-                }
-            } catch (e: Exception) {
-                withContext(Dispatchers.Main) {
-                    Log.e("TestConnection", "Exception occurred: ${e.message}", e)
-                }
-            }
-        }
-    }
-
-    private fun testDataStore() {
         lifecycleScope.launch {
-            val testToken = "sample_jwt_token"
-
-            // Save the token
-            DataStoreManager.saveToken(applicationContext, testToken)
-            Log.d("DataStoreTest", "Token saved: $testToken")
-
-            // Retrieve the token
-            val retrievedToken = DataStoreManager.getToken(applicationContext).first()
-            Log.d("DataStoreTest", "Token retrieved: $retrievedToken")
-
-            // Clear the token
-            DataStoreManager.clearToken(applicationContext)
-            Log.d("DataStoreTest", "Token cleared")
-
-            // Check if the token is really removed
-            val emptyToken = DataStoreManager.getToken(applicationContext).first()
-            Log.d("DataStoreTest", "Token after clearing: $emptyToken")
+            DataStoreManager.testDataStore(applicationContext)
         }
+
     }
+
 }
