+++ app/src/main/java/com/copay/app/repository/ExpenseRepository.kt
@@ -5,7 +5,7 @@ import android.util.Log
 import com.copay.app.dto.expense.request.GetExpenseRequestDTO
 import com.copay.app.dto.expense.response.GetExpenseResponseDTO
 import com.copay.app.service.ExpenseService
-import com.copay.app.utils.TokenUtils
+import com.copay.app.utils.DataStoreManager
 import com.copay.app.utils.state.ExpenseState
 import com.google.gson.Gson
 import com.google.gson.JsonObject
@@ -27,7 +27,7 @@ class ExpenseRepository(private val expenseService: ExpenseService) {
 
         val request = GetExpenseRequestDTO(groupId = groupId)
 
-        val token = TokenUtils.getFormattedToken(context)
+        val token = DataStoreManager.getFormattedToken(context)
 
         return handleApiResponse(context) { expenseService.getExpenses(request, token) }
     }
